requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
decorator==5.1.1
Deprecated==1.2.14
docker-pycreds==0.4.0
executing==1.2.0
filelock==3.12.4
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.36
humanize==4.8.0
idna==3.4
ipdb==0.13.13
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.4.0
mpmath==1.3.0
munch==4.0.0
networkx==3.1
nibabel==5.1.0
numpy==1.22.0
packaging==23.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
rich==13.5.3
scikit-learn==1.2.2
scipy==1.11.2
sentry-sdk==1.31.0
setproctitle==1.3.2
shellingham==1.5.3
SimpleITK==2.3.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchio==0.19.1
tqdm==4.66.1
traitlets==5.10.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.4
wandb==0.15.10
wcwidth==0.2.6
wrapt==1.15.0
